bugfix/false-positive-local-variable-early-declaration(#488) ### What's done:   * Fixed bugs
