+++ diktat-rules/src/test/resources/test/smoke/src/main/kotlin/Example4Expected.kt
@@ -1,0 +1,12 @@
+package org.cqfn.diktat
+
+fun `method name incorrect, part 4`() {
+    val code = """
+                  class TestPackageName {
+                    fun methODTREE(): String {
+                    }
+                  }
+                """.trimIndent()
+    lintMethod(code, LintError(2, 7, ruleId, "${FUNCTION_NAME_INCORRECT_CASE.warnText()} methODTREE", true))
+}
+
